Generalize transform methods
